Do not multiply an enumerator by a float
GROUP_COUNT is an unscoped enumerator, and C++20 deprecates arithmetic
between such an enumerator and a floating-point type. GCC 15 reports it,
GCC 14 does not, which is why the local build looked clean while CI failed.

Casting to float says what was meant: use the number of groups as a count.

+++ src/draw_batteryinfo_bi.cpp
@@ -1570,7 +1570,7 @@ void draw_batteryinfo_bi::drawSettingsTab(ID2D1HwndRenderTarget *pRT, init_dwrit
         pushHit(D2D1::RectF(L, ry, L + RAIL_W, ry + RAIL_ITEM_H), HIT_RAIL, g, nullptr);
     }
 
-    float railBottom = railTop + GROUP_COUNT * (RAIL_ITEM_H + RAIL_GAP);
+    float railBottom = railTop + (float)GROUP_COUNT * (RAIL_ITEM_H + RAIL_GAP);
 
     D2D1_SIZE_F panel = measureOverlayPanel(ov);
